requirements.txt
boto3==1.34.151
langchain==0.2.11
pandas==2.2.2
pymongo==4.8.0
python-dotenv==1.0.1